tests/wpt/mozilla/tests/mozilla/document_characterSet_default.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
            test(function() {
                // FIXME: https://github.com/mozilla-servo/libhubbub/issues/5
                assert_equals(document.characterSet, "UTF-8", "test1-0, default characterSet");
            });
        </script>
    </head>
</html>
